добавил редирект на страницу, с которой был сделан запрос при добавлении или удалении товара в корзине, добавил возможность удалять товары из корзины по одному

+++ source/webapp/views/product_in_cart.py
@@ -1,9 +1,7 @@
-from django.shortcuts import render, get_object_or_404, redirect
-from django.urls import reverse
-from django.views.generic import ListView, CreateView, DetailView, UpdateView, DeleteView, View
+from django.shortcuts import get_object_or_404, redirect
+from django.views.generic import ListView, View
 
-from webapp.forms import ProductForm, SearchForm
-from webapp.models import Product, Category, ProductInCart
+from webapp.models import Product, ProductInCart
 
 
 class ProductsInCartListView(ListView):
@@ -26,22 +24,21 @@ class AddProductInCartView(View):
         product = get_object_or_404(Product, pk=kwargs.get('pk'))
         if product.remain > 0:
             product_in_cart, created = ProductInCart.objects.get_or_create(product=product, defaults={'quantity': 1})
-            print(product_in_cart)
             if not created:
                 if product_in_cart.quantity < product.remain:
                     product_in_cart.quantity += 1
                     product_in_cart.save()
-                    return redirect('products')
-                else:
-                    return redirect('products')
-            else:
-                return redirect('products')
+            return redirect(request.META['HTTP_REFERER'])
+        return redirect(request.META['HTTP_REFERER'])
 
 
 class DeleteProductInCartView(View):
 
     def get(self, request, *args, **kwargs):
         product_in_cart = get_object_or_404(ProductInCart, pk=kwargs.get('pk'))
-        product_in_cart.delete()
-        return redirect('cart')
-
+        if product_in_cart.quantity > 1:
+            product_in_cart.quantity -= 1
+            product_in_cart.save()
+        else:
+            product_in_cart.delete()
+        return redirect(request.META['HTTP_REFERER'])
